Performance improvements
Generating the DIF was much slower than the old GUI, due to updating
the statusbar. I now run the DIF generation in a separate thread, which
speeds up thinks considerably. It is still around 20-25% slower than
the other GUI (which does not show any progress). Not really that much.
A 40GB MRI data set runs in 80 seconds, compared to 65 seconds in the
old GUI. We could make showing the progress an option in the GUI, so it
can be switched off if preferred.